app1/backend/requirements.txt
Flask==3.0.3
google-cloud-storage==2.18.2
joblib==1.4.2
Pillow==10.4.0
numpy==1.26.4
tensorflow==2.17.0
scikit-learn==1.5.1